third_party/blink/web_tests/external/wpt/css/css-transitions/parsing/transition-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Default transition value
  assert_equals: expected "all" but got "all 0s ease 0s"
[FAIL] Property transition value '1s'
  assert_equals: expected "1s" but got "all 1s ease 0s"
[FAIL] Property transition value 'cubic-bezier(0, -2, 1, 3)'
  assert_equals: expected "cubic-bezier(0, -2, 1, 3)" but got "all 0s cubic-bezier(0, -2, 1, 3) 0s"
[FAIL] Property transition value '1s -3s'
  assert_equals: expected "1s -3s" but got "all 1s ease -3s"
[FAIL] Property transition value 'none'
  assert_equals: expected "none" but got "none 0s ease 0s"
[FAIL] Property transition value 'top'
  assert_equals: expected "top" but got "top 0s ease 0s"
[FAIL] Property transition value '1s -3s, cubic-bezier(0, -2, 1, 3) top'
  assert_equals: expected "1s -3s, top cubic-bezier(0, -2, 1, 3)" but got "all 1s ease -3s, top 0s cubic-bezier(0, -2, 1, 3) 0s"
[FAIL] Property transition value 'all, all'
  assert_equals: expected "all, all" but got "all 0s ease 0s, all 0s ease 0s"
[FAIL] Transition with a delay but no duration
  assert_equals: expected "0s 1s" but got "all 0s ease 1s"
Harness: the test ran to completion.
